// apps/web/src/lib/api/generated/main/model/killsControllerGetUserNpcKillsParams.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { KillsControllerGetUserNpcKillsNpcTypesItem } from './killsControllerGetUserNpcKillsNpcTypesItem';
import type { KillsControllerGetUserNpcKillsSortBy } from './killsControllerGetUserNpcKillsSortBy';
import type { KillsControllerGetUserNpcKillsSortOrder } from './killsControllerGetUserNpcKillsSortOrder';

export type KillsControllerGetUserNpcKillsParams = {
npcTypes?: KillsControllerGetUserNpcKillsNpcTypesItem[];
world?: string;
search?: string;
/**
 * @minimum 0
 * @maximum 9007199254740991
 */
cursor?: number;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
sortOrder?: KillsControllerGetUserNpcKillsSortOrder;
sortBy?: KillsControllerGetUserNpcKillsSortBy;
/**
 * @minimum 0
 * @maximum 500
 */
minLvl?: number;
/**
 * @minimum 0
 * @maximum 500
 */
maxLvl?: number;
};